django-backend/requirements.txt
asgiref
Django
djangorestframework
setuptools
sqlparse
typing_extensions
wheel
psycopg2-binary
newsapi-python
cohere
transformers
torch
accelerate
sentencepiece
protobuf
pgvector
google-generativeai
google-genai
streamlit
pandas
numpy
kafka-python